+++ include/readseek_impl.h
@@ -7,7 +7,7 @@ public:
   ReadSeekBuf(const std::vector<uint8_t>& buf): Buf(buf), Pos(0) {}
   virtual ~ReadSeekBuf() {}
 
-  virtual ssize_t read(size_t len, std::vector<uint8_t>& buf) override;
+  virtual int64_t read(size_t len, std::vector<uint8_t>& buf) override;
 
   virtual size_t tellg() const override { return Pos; }
   virtual size_t seek(size_t pos) override { return (Pos = (pos < Buf.size() ? pos: Buf.size())); }
@@ -27,7 +27,7 @@ public:
   ReadSeekFile(std::shared_ptr<FILE> fileptr);
   virtual ~ReadSeekFile() {}
 
-  virtual ssize_t read(size_t len, std::vector<uint8_t>& buf) override;
+  virtual int64_t read(size_t len, std::vector<uint8_t>& buf) override;
 
   virtual size_t tellg() const override;
   virtual size_t seek(size_t pos) override;
@@ -49,7 +49,7 @@ public:
   ReadSeekTSK(TSK_FS_ATTR* attr); // does not take ownership
   virtual ~ReadSeekTSK() {}
 
-  virtual ssize_t read(size_t len, std::vector<uint8_t>& buf) override;
+  virtual int64_t read(size_t len, std::vector<uint8_t>& buf) override;
 
   virtual size_t tellg() const override { return Pos; }
   virtual size_t seek(size_t pos) override;
